Se modifica el archivo de inicio
